+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -148,6 +148,11 @@ New checks
   Warns if anything other than main(), operator new/delete, namespaces, or
   extern "C" declarations are present at global scope.
 
+- New :doc:`bsl-nonPOD-class-def
+  <clang-tidy/checks/bsl-nonPOD-class-def>` check.
+
+  FIXME: add release notes.
+
 - New :doc:`bsl-op-bitwise-operands
   <clang-tidy/checks/bsl-op-bitwise-operands>` check.
 
@@ -382,6 +387,9 @@ Changes in existing checks
 
 Renamed checks
 ^^^^^^^^^^^^^^
+
+- The 'bsl-nonPOD-class-def' check was renamed to :doc:`bsl-non-pod-classdef
+  <clang-tidy/checks/bsl-non-pod-classdef>`
 
 - The 'fuchsia-restrict-system-headers' check was renamed to :doc:`portability-restrict-system-includes
   <clang-tidy/checks/portability-restrict-system-includes>`